Fix missing int64 in cmd caller
Description:
The map that is initialized had int for key (I missed it when I
changed everything over to int64).

Note: Minor update

+++ pipeline/backend/cmd/backend.go
@@ -15,6 +15,6 @@ func main() {
 
 	backendServer := backend.Server{}
 	backendServer.Repos = new(backend.ActiveRepos)
-	backendServer.Repos.Actives = make(map[int]*backend.ArchRepo)
+	backendServer.Repos.Actives = make(map[int64]*backend.ArchRepo)
 	backendServer.Start()
 }
